+++ zrml/neo-swaps/src/liquidity_tree/mod.rs
@@ -15,10 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
 
-mod fee_distribution;
-mod pool;
-mod solo_lp;
-
-pub(crate) use fee_distribution::*;
-pub(crate) use pool::*;
-pub(crate) use solo_lp::*;
+mod macros;
+mod tests;
+pub(crate) mod traits;
+pub(crate) mod types;
