+++ zrml/neo-swaps/src/liquidity_tree/tests/shares_of.rs
@@ -15,10 +15,15 @@
 // You should have received a copy of the GNU General Public License
 // along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
 
-mod fee_distribution;
-mod pool;
-mod solo_lp;
+use super::*;
 
-pub(crate) use fee_distribution::*;
-pub(crate) use pool::*;
-pub(crate) use solo_lp::*;
+#[test]
+fn shares_of_works() {
+    let tree = utility::create_test_tree();
+    assert_eq!(tree.shares_of(&3).unwrap(), _1);
+    assert_eq!(tree.shares_of(&9).unwrap(), _3);
+    assert_eq!(tree.shares_of(&5).unwrap(), _3);
+    assert_eq!(tree.shares_of(&7).unwrap(), _1);
+    assert_eq!(tree.shares_of(&6).unwrap(), _12);
+    assert_eq!(tree.shares_of(&8).unwrap(), _4);
+}
