+++ zrml/neo-swaps/src/liquidity_tree/traits/mod.rs
@@ -15,10 +15,6 @@
 // You should have received a copy of the GNU General Public License
 // along with Zeitgeist. If not, see <https://www.gnu.org/licenses/>.
 
-mod fee_distribution;
-mod pool;
-mod solo_lp;
+pub(crate) mod liquidity_tree_helper;
 
-pub(crate) use fee_distribution::*;
-pub(crate) use pool::*;
-pub(crate) use solo_lp::*;
+pub(crate) use liquidity_tree_helper::*;
